Fix M114 to show fractional units

+++ src/com_interpreter.c
@@ -565,7 +565,7 @@ void process_commands()
 		//printf("X:%d Y:%d Z:%d E:%d\r\n",(int)pa.axis_steps_per_unit[0],(int)pa.axis_steps_per_unit[1],(int)pa.axis_steps_per_unit[2],(int)pa.axis_steps_per_unit[3]);
         break;
 	  case 114: // M114 Display current position
-		usb_printf("X:%d Y:%d Z:%d E:%d",(int)current_position[0],(int)current_position[1],(int)current_position[2],(int)current_position[3]);
+		usb_printf("X:%f Y:%f Z:%f E:%f",current_position[0],current_position[1],current_position[2],current_position[3]);
         break;
       case 115: // M115
         usb_printf("FIRMWARE_NAME: Sprinter 4pi PROTOCOL_VERSION:1.0 MACHINE_TYPE:Prusa EXTRUDER_COUNT:%d\r\n",MAX_EXTRUDER);
